+++ src/main/java/com/bookiply/interview/assignment/services/IGeoLocationService.java
@@ -3,6 +3,5 @@ package com.bookiply.interview.assignment.services;
 import com.bookiply.interview.assignment.models.GeoPoint;
 
 public interface IGeoLocationService {
-
-    double distance(GeoPoint startP, GeoPoint endP );
+    long distance(GeoPoint startP, GeoPoint endP);
 }
